+++ src/main/java/io/bux/matchingengine/domain/TradeModel.java
@@ -3,10 +3,12 @@ package io.bux.matchingengine.domain;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class TradeModel {
     private final Long orderId;
-    private final Double price;
-    private final Double amount;
+    private final BigDecimal price;
+    private final BigDecimal amount;
 }
