+++ src/main/java/io/bux/matchingengine/dto/OrderResponse.java
@@ -2,6 +2,7 @@ package io.bux.matchingengine.dto;
 
 import io.bux.matchingengine.enums.Direction;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 
@@ -9,11 +10,11 @@ public record OrderResponse(Long id,
                             //@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSXXX")
                             Instant timestamp,
                             String asset,
-                            Double price,
-                            Double amount,
+                            BigDecimal price,
+                            BigDecimal amount,
                             Direction direction,
                             List<TradeResponse> trades,
-                            Double pendingAmount) {
+                            BigDecimal pendingAmount) {
 
     @Override
     public String toString() {
